Move txn stuff to drivers and eval second txn call
Moved the calls to `begin_work`, `commit`, and `rollback` to the drivers, as
ribasushi tells me that there are some problems with those implemented by
DBD::Sybase, and there are currently workarounds in a branch of DBIx::Class.
DBIx::Connection now calls only one DBI method directly, `disconnect`. All
others are handled by the driver classes.

Also realized that the second call to the sub in `txn_do` should be
rolled-back in the event of a failure, so made that change, too, and added
tests for it.