Added notes for exception-handling to-do.
I think it'd be best done by passing another code ref to `run()` and friends,
which would work nicely with Try::Tiny.

Also removed the to-do about exception objects. There is no place that the
code dies with a string, it simply rethrows exceptions from the blocks. And
because those exceptions end in a newline (when created via `die $string`,
becaus Perl adds a newline after the file and line number), it's context
should always be correct.