Set `AutoInactiveDestroy` with DBI 1.614 and up.
Set `AutoInactiveDestroy` on connect with DBI 1.614 and higher, unless it is
explicitly set in the attributes. This makes things even safer in a forking
environment, preventing a parent process from getting disconnected when a
child exits without using the connection. The reports from Peter Rabbitson and
Aran Deltac and subsequent discussion with Tim Bunce led to the addition of
this attribute in DBI 1.614, which is now the recommended version of DBI.