Fix bug where catch blocks did not propagate exceptions.
If you threw an exception from a catch block, it would not be detectable
from outside:

    eval {
        $conn->run(sub { die 'WTF' }, catch => sub { die 'OMG!' });
    };
    if (my $err = $@) {
        say "We got an error: $@\n";
    }

In this example, the `if` block would never be called, because even though the
catch block threw an exception, $@ was not set. This could be terribly
annoying, as you can imagine. I was being a bit too clever about localizing
`$@`, but did so with the scope much too broad. This patch uses a much tighter
scope to localize `$@`, so now it should propagate properly everywhere.