Default RaiseError to true.
It's important for the execution methods that the database handle raise
exceptions on errors so that they can be properly handled. This is especially
true for C<txn()> to get automatic rollback on failure. So set `RaiseError`
to true if neither it nor `HandleError` has been specified. And document
how important it is that these attributes be set correctly.

While at it, document that DBIx::Connector also chagnes the default value of
`AutoInactiveDestroy`.