Don't clear CachedKids on disconnect.
I originally copied this code from DBIx::Class::Storage::DBI, but issue #39
reports that it's not only unnecessary on PostgreSQL, but can lead to noisy
log output as prepared statements are deallocated. Discussion with the
DBIx::Class folks suggests that the code was originally added when Sybase ASA
and SQLAnywhere support were added. So it's much more likely to be necessary
only for specific database back ends or drivers. If we ever run into it, we
can add a `_disconnect` method to DBIx::Connector::Driver and override it in
the relevant driver-specific subclass. But for now, I don't believe it's
necessary at all.

Resolves #39 and closes #40.