Don't require autoincrement column in test.
It just makes it difficult for Firebird, and likely other databases, and is
just not necessary for the purposes of the test. Instead, use an explicit ID
value. This allows us to avoid all the generator stuff on Firebird, as I ran
into an issue where it would die on `DROP GENERATOR`, because the first time I
ran it there was no generator to drop!x