bump mininum Perl to 5.8.1
DBIx::Connector has always claimed to support Perl 5.6.2.

However it has always also required DBI 1.605. The last version of DBI
which works on Perl 5.6.2 is 1.604 – in fact, the last version to have
only PASSes on 5.6 on CPAN Testers is 1.603.

So DBIx::Connector has never worked on 5.6.2. The simplest resolution
for this situation is to properly document the status quo by bumping the
minimum Perl requirement to 5.8.1 (where DBI is solidly green even today
(at version 1.643 as of this writing)).